Fix GitVersion configuration
Workflows use continuous deployment to automatically build
new packages on changes. This mean ContinuousDeployment
mode must be used on all branches with automatic workflows
to provide appropriate semantic versions for the packages to not
conflict in the destination feeds.

Only the main branch is manually triggered, so set all configuration
to ContinuousDeployment and only override main.